Scripts/requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0

# Computer Vision
opencv-python>=4.5.0
scikit-image>=0.19.0

# Machine Learning Utils
scikit-learn>=1.0.0
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0

# Dataset Download
gdown>=4.6.0

# Model Export
onnx>=1.12.0
onnxruntime>=1.13.0

# OAK-D Camera (Optional - for deployment only)
depthai>=2.20.0

# Blob Conversion (Optional - for OAK-D)
blobconverter>=1.4.0

# Progress Bars
tqdm>=4.65.0